Move CLI tests
