+++ testsuite-core/domain/src/test/java/org/jboss/as/test/integration/domain/management/cli/BasicOpsTestCase.java
